requirements.txt
streamlit
pymysql
python-dotenv
mysql-connector-python 
bcrypt
PyJWT==2.6.0
